Work with a very restricted object mapper
The sample code relied on a number of internal features of Jackson
that are turned on by default. If those are turned off (e.g. if there
is no control over the object mapper that is used), the sample code
actually does not work.

This tweaks the demo code a bit so that even with a very restricted
object mapper, the sample code still works.